+++ WebMvcPluginTour/WebMvcPluginTour/Controllers/AdminController.cs
@@ -175,7 +175,7 @@ namespace WebMvcPluginTour.Controllers
                         || !CoreHelper.GetParameter(out var jsonName, body, "Name",
                             JTokenType.String, ref responseModel)
                         || !CoreHelper.GetParameter(out var jsonPrice, body, "Price",
-                            JTokenType.Integer, ref responseModel)
+                            JTokenType.Float, ref responseModel)
                         || !CoreHelper.GetParameter(out var jsonServices, body, "Services",
                             JTokenType.Array, ref responseModel, isNullable: true))
                     {
@@ -197,7 +197,7 @@ namespace WebMvcPluginTour.Controllers
                     _ = int.TryParse(jsonMaxMember?.ToString(), out var maxMember);
                     _ = int.TryParse(jsonTotalDay?.ToString(), out var totalDay);
                     _ = int.TryParse(jsonTotalNight?.ToString(), out var totalNight);
-                    _ = int.TryParse(jsonPrice?.ToString(), out var price);
+                    _ = float.TryParse(jsonPrice?.ToString(), out var price);
                     var serviceIds = jsonServices != null
                         ? JsonConvert.DeserializeObject<string[]>(jsonServices.ToString())
                         : null;
@@ -270,7 +270,7 @@ namespace WebMvcPluginTour.Controllers
                         || !CoreHelper.GetParameter(out var jsonTotalNight, body, "TotalNight",
                             JTokenType.Integer, ref responseModel, isNullable: true)
                         || !CoreHelper.GetParameter(out var jsonPrice, body, "Price",
-                            JTokenType.Integer, ref responseModel, isNullable: true)
+                            JTokenType.Float, ref responseModel, isNullable: true)
                         || !CoreHelper.GetParameter(out var jsonName, body, "Name",
                             JTokenType.String, ref responseModel, isNullable: true)
                         || !CoreHelper.GetParameter(out var jsonTimelines, body, "Timelines",
@@ -293,7 +293,7 @@ namespace WebMvcPluginTour.Controllers
                     var isMaxMemberParse = int.TryParse(jsonMaxMember?.ToString(), out var maxMember);
                     var isTotalDayParsed = int.TryParse(jsonTotalDay?.ToString(), out var totalDay);
                     var isTotalNightParsed = int.TryParse(jsonTotalNight?.ToString(), out var totalNight);
-                    var isPriceParsed = int.TryParse(jsonPrice?.ToString(), out var price);
+                    var isPriceParsed = float.TryParse(jsonPrice?.ToString(), out var price);
                     var services = jsonServices != null
                         ? JsonConvert.DeserializeObject<string[]>(jsonServices.ToString())
                         : null;
@@ -311,7 +311,7 @@ namespace WebMvcPluginTour.Controllers
                         endDay: isEndDayParse ? endDate : (DateTime?) null,
                         totalDay: isTotalDayParsed ? totalDay : (int?) null,
                         totalNight: isTotalNightParsed ? totalNight: (int?) null,
-                        price: isPriceParsed ? price: (int?) null,
+                        price: isPriceParsed ? price: (float?) null,
                         maxMember: isMaxMemberParse ? maxMember : (int?) null,
                         services: services
                     );
